add try&catch

+++ client-react/src/components/body/msDetails/ShowMSDetails.jsx
@@ -8,6 +8,7 @@ import { Button } from "primereact/button";
 import FormMSDetails from "./FormMSDetails";
 import { useSelector } from "react-redux";
 import { Toolbar } from "primereact/toolbar";
+import { format } from 'date-fns';
 
 const ShowMSDetails = () => {
     const { token } = useSelector((state) => state.token);
@@ -16,10 +17,14 @@ const ShowMSDetails = () => {
     const printRef = useRef(null); // Ref for the print container
     const [MSDetails, setMSDetails] = useState([]);
     const [MSDetail, setMSDetail] = useState({});
-    
+    const [enableAdd, setEnableAdd] = useState(true)
     const getAllMSDetails = async () => {
         const res = await axios.get('http://localhost:1111/api/monthlyScholarshipDetails',
             { headers: { Authorization: `Bearer ${token}` } });
+        const addedMsdetails = await axios.get('http://localhost:1111/api/monthlyScholarshipDetails/thisMonth',
+            { headers: { Authorization: `Bearer ${token}` } });
+        if (addedMsdetails.data !== "")
+            setEnableAdd(false)
         setMSDetails(res.data);
     };
 
@@ -32,23 +37,43 @@ const ShowMSDetails = () => {
     };
 
     const deleteMSDetails = async (rowData) => {
-        if (window.confirm("Are you sure you want to delete this record?")) {
-            const res = await axios.delete(`http://localhost:1111/api/monthlyScholarshipDetails/${rowData._id}`,
-                { headers: { Authorization: `Bearer ${token}` } });
-            getAllMSDetails();
+
+        const newDate = new Date(rowData.date)
+        console.log(newDate,"newDate");
+        if (newDate.getMonth() === new Date().getMonth() && newDate.getFullYear() === new Date().getFullYear()) {
+            if (window.confirm("Are you sure you want to delete this record?")) {
+                {
+                    try{
+                    const res = await axios.delete(`http://localhost:1111/api/monthlyScholarshipDetails/${rowData._id}`,
+                        { headers: { Authorization: `Bearer ${token}` } });}
+                        catch (err) {
+                            console.error(err);
+                        }
+                    getAllMSDetails();
+                    setEnableAdd(true)
+                }
+            }
+        } else {
+            alert("You can't delete monthly-scholarship-details from last month")
+            return
         }
     };
 
     const updateButton = (rowData) => {
         return (
-          
             <div className="flex align-items-center gap-2">
 
             <Button  icon="pi pi-pencil" onClick={() => {
-                setMSDetail(rowData);
-                setVisible(true);
-            }} className="p-button-rounded"></Button>            </div>
-
+                const newDate = new Date(rowData.date)
+                if (newDate.getMonth() === new Date().getMonth() && newDate.getFullYear() === new Date().getFullYear()) {
+                    setMSDetail(rowData);
+                    setVisible(true);
+                }
+                else {
+                    alert("You can't update monthly-scholarship-details from last month")
+                    return
+                }
+            }} className="p-button-rounded"></Button></div>
         );
     };
 
@@ -77,22 +102,27 @@ const ShowMSDetails = () => {
 
     const startContent = (
         <React.Fragment>
-            <Button icon="pi pi-plus" className="mr-2" onClick={createMSDatalis} />
+            <Button icon="pi pi-plus" className="mr-2" onClick={createMSDatalis} visible={enableAdd} />
             <Button icon="pi pi-print" className="mr-2" onClick={handlePrint} />
-            <Button label="Export" icon="pi pi-upload" className="p-button-help" onClick={exportCSV} />
+            <Button label="Export" icon="pi pi-download" iconPos="right" className="p-button-help" onClick={exportCSV} />
         </React.Fragment>
     );
+    const dateBodyTemplate = (rowData) => {
+        if (rowData.date)
+            return format(rowData.date, 'dd/MM/yyyy')
+        return ""
+    };
 
     return (
         <>
             <div className="card">
                 <Toolbar center={startContent} />
             </div>
-            <div  className="card">
+            <div className="card">
                 <DataTable ref={ms} value={MSDetails} tableStyle={{ minWidth: '50rem' }}>
                     <Column field="sumPerHour" header="SumPerHour"></Column>
                     <Column field="MaximumNumberOfHours" header="MaximumNumberOfHours"></Column>
-                    <Column field="date" header="Date"></Column>
+                    <Column field="date" header="Date" body={dateBodyTemplate}></Column>
                     <Column header="DELETE" body={deleteButton}></Column>
                     <Column header="UPDATE" body={updateButton}></Column>
                 </DataTable>
